Fixing behaviour where composer is re-installed even if already present
Previously this role did nothing to check if composer was already installed. It now checks and if something exists at the path composer is symlinked to (i.e. /usr/bin/composer) download and installation will be skipped.

This means the latest version won't always be installed but and would require composer to first be removed before a new version would be installed. This is probably sensible so there's consistency with the possible future addition of configuration option to first remove composer so a newer version would be installed.